requirements.txt
requests==2.22.0
python-digitalocean==1.16.0
pydig==0.3.0